+++ expense-app/package.json
@@ -15,6 +15,7 @@
     "@types/react": "^18.0.10",
     "@types/react-dom": "^18.0.5",
     "babel-plugin-macros": "^3.1.0",
+    "bootstrap": "^5.1.3",
     "react": "^18.1.0",
     "react-chartjs-2": "^4.1.0",
     "react-dom": "^18.1.0",
