+++ expense-app/src/components/Tabs/HomeTab.tsx
@@ -1,7 +1,12 @@
 import { faPlus } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React, { useState } from "react";
-import { tableData, CategoriesColumns, TransactionColumns, catData } from "../../data";
+import {
+  tableData,
+  CategoriesColumns,
+  TransactionColumns,
+  catData,
+} from "../../data/data";
 import Modal from "../modal/Modal";
 
 import Table from "../Table/Table";
@@ -11,7 +16,6 @@ const HomeTab = () => {
   const [toggleState, setToggleState] = useState(1);
   const [openTransactionModal, setOpenTransactionModal] = useState(false);
   const [openCategoryModal, setOpenCategoryModal] = useState(false);
-  
 
   const toggleTab = (index: number) => {
     setToggleState(index);
@@ -24,13 +28,13 @@ const HomeTab = () => {
           className={toggleState === 1 ? "tabs active-tabs" : "tabs"}
           onClick={() => toggleTab(1)}
         >
-         Transactions
+          Transactions
         </button>
         <button
           className={toggleState === 2 ? "tabs active-tabs" : "tabs"}
           onClick={() => toggleTab(2)}
         >
-         Categories
+          Categories
         </button>
       </div>
 
@@ -38,21 +42,43 @@ const HomeTab = () => {
         <div
           className={toggleState === 1 ? "content  active-content" : "content"}
         >
-          <div className="header-button"><h2>Transactions</h2>
-          <button className="button-add" onClick={() => setOpenTransactionModal(true)}><FontAwesomeIcon icon={faPlus}></FontAwesomeIcon></button>
-           </div>
-          <Table tableData={tableData} columns={TransactionColumns} del={false}/>
-          {openTransactionModal && <Modal modal={'Add Transaction'} updateModal={setOpenTransactionModal} />}
+          <div className="header-button">
+            <h2>Transactions</h2>
+            <button
+              className="button-add"
+              onClick={() => setOpenTransactionModal(true)}
+            >
+              <FontAwesomeIcon icon={faPlus}></FontAwesomeIcon>
+            </button>
+          </div>
+          <Table
+            tableData={tableData}
+            columns={TransactionColumns}
+            del={false}
+          />
+          {openTransactionModal && (
+            <Modal
+              modal={"Add Transaction"}
+              updateModal={setOpenTransactionModal}
+            />
+          )}
         </div>
         <div
           className={toggleState === 2 ? "content  active-content" : "content"}
         >
           <div className="header-button">
-            <h2>Categories</h2> 
-            <button className="button-add" onClick={() => setOpenCategoryModal(true)}><FontAwesomeIcon icon={faPlus} ></FontAwesomeIcon></button>
-            </div>
-          <Table tableData={catData} columns={CategoriesColumns} del={true}/>
-         {openCategoryModal && <Modal modal={'Add Category'} updateModal={setOpenCategoryModal} />} 
+            <h2>Categories</h2>
+            <button
+              className="button-add"
+              onClick={() => setOpenCategoryModal(true)}
+            >
+              <FontAwesomeIcon icon={faPlus}></FontAwesomeIcon>
+            </button>
+          </div>
+          <Table tableData={catData} columns={CategoriesColumns} del={true} />
+          {openCategoryModal && (
+            <Modal modal={"Add Category"} updateModal={setOpenCategoryModal} />
+          )}
         </div>
       </div>
     </div>
